Change deployment setup
