+++ .gitignore
@@ -3,3 +3,4 @@ node_modules
 lib
 build/
 **/__pycache__
+backend/ontario-web/public/
