--- backend/ontario-web/public/index.html
@@ -1,5 +1,0 @@
-<html>
-<body>
-<p>Test worked!</p>
-</body>
-</html>
