--- deploy/.env
@@ -1,8 +1,0 @@
-# GNU AGPL v3.0
-# Written by John Nunley
-
-POSTGRES_USER=jtnunley
-POSTGRES_PASSWORD=password123
-POSTGRES_DB=pictonode
-POSTGRES_HOST=localhost
-POSTGRES_PORT=5432
